refactor to move logic into text switch link component
